+++ src/wally/github.ts
@@ -1,5 +1,7 @@
 import * as vscode from "vscode";
 
+import { pMemoizeDecorator } from "p-memoize";
+
 import { Octokit } from "@octokit/rest";
 
 import { RequestError } from "@octokit/request-error";
@@ -233,6 +235,7 @@ export class WallyGithubHelper {
 		}
 	}
 	
+	@pMemoizeDecorator()
 	private async getGitHubTree(sha: string) {
 		if (this.registryUser && this.registryRepo) {
 			return this.kit.git.getTree({
@@ -255,6 +258,7 @@ export class WallyGithubHelper {
 		return null;
 	}
 	
+	@pMemoizeDecorator()
 	private async getGitHubBlob(sha: string) {
 		if (this.registryUser && this.registryRepo) {
 			return this.kit.git.getBlob({
@@ -287,6 +291,7 @@ export class WallyGithubHelper {
 	
 	
 	
+	@pMemoizeDecorator()
 	private async getRegistryTree(): Promise<WallyGithubRegistryTree | null> {
 		if (this.registryUser && this.registryRepo) {
 			// Check for cached tree
@@ -333,6 +338,7 @@ export class WallyGithubHelper {
 		return null;
 	}
 	
+	@pMemoizeDecorator()
 	private async getRegistryAuthor(authorName: string): Promise<WallyGithubRegistryAuthor | null> {
 		// Check for cached authors
 		const lowered = authorName.toLowerCase();
@@ -378,6 +384,7 @@ export class WallyGithubHelper {
 		return null;
 	}
 	
+	@pMemoizeDecorator()
 	private async getRegistryPackage(authorName: string, packageName: string): Promise<WallyGithubRegistryPackage | null> {
 		// Check for cached packages
 		const loweredAuthor = authorName.toLowerCase();
@@ -431,6 +438,7 @@ export class WallyGithubHelper {
 		return null;
 	}
 	
+	@pMemoizeDecorator()
 	private async getRegistryConfig(): Promise<WallyGithubRegistryConfig | null> {
 		const tree = await this.getRegistryTree();
 		if (tree && this.registryUser && this.registryRepo) {